+++ Engine/Window.h
@@ -33,7 +33,7 @@ public:
 
     glm::vec2 getDisplaySize() const;
 
-    glm::vec2 getDrawableSize() const;
+    glm::ivec2 getDrawableSize() const;
 
     const char *getClipboardText() const;
